+++ pkg/odo/cli/project/set.go
@@ -68,7 +68,9 @@ func (pso *ProjectSetOptions) Validate() (err error) {
 
 // Run runs the project set command
 func (pso *ProjectSetOptions) Run(cmd *cobra.Command) (err error) {
-	scontext.SetClusterType(cmd.Context(), pso.Client)
+	if scontext.GetTelemetryStatus(cmd.Context()) {
+		scontext.SetClusterType(cmd.Context(), pso.Client)
+	}
 	current := pso.Project
 	err = project.SetCurrent(pso.Context, pso.projectName)
 	if err != nil {